Image_captioning/requirements.txt
einops==0.4.1
gensim==4.2.0
h5py==3.7.0
joblib==1.1.0
lmdb==1.3.0
lmdbdict==0.2.2
opencv-python==4.6.0.66
Pillow==5.1.0
pycocotools
pyemd==0.5.1
PyYAML==6.0
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.1.0
six
tensorboard==2.10.0
tensorboardX==2.5.1
timm==0.6.7
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
yacs==0.1.8
image
matplotlib
